requirements.txt
django==3.0.7
channels==2.4.0
channels_redis==2.4.2
django-jalali-date==0.3.1
djangorestframework==3.11.0
django-cors-headers==3.4.0